fix(timer_manager): get stats of n days before
1. When we issue the command: `> pt stats 1` we should get the stats of
yesterday. But as the above command was getting normalized as
`{command: "STATS", argument: "1"}` the timer_manager manager thought
that we were trying to get the stats of a specific date and was throwing
error as "1" is not a valid date string. So the commandObjectNormalizer
needed some mechanism to convert that "1" into a number and that's what
I've added in this commit. Now we can specify an $exec method in command
schemas for arguments.
1. Also added default value for arguments in schemas