feat: delete all modules and migrate to Typescript
Sorry, I really don't know if this commit should be in the feat
category! I couldn't any suitable category.
In this commit I've deleted all the modules in the src and tests folder
because I want to rewrite this application with a completely different
architecture. InshaAllah it'll include a lot of new features and
capabilities that was not intended in the previous version.