feat(countdown-time): improve time error handling
Previously, if the difference between the lastTick and currentTick were
less than TICK_INTERVAL_MS the timer emitted the "err:time_decrement"
event and paused the timer. But I've found that in real life sometimes
the diff can be less than TICK_INTERVAL_MS. I've tested a timer with
TICK_INTERVAL_MS set to 1000ms and sometimes the tickDiff was 999
(weird!) and that paused the timer. So I've added an option
"MIN_ALLOWED_TICK_DIFF_MS" for fine grain control.