refactor(src/data-access): remove the overengineered subprocess db
While a timer is running, if we change the system time and the
difference between the current and last tick becomes greater than 1s,
then the timer pauses itself and emit a "err:wake_up_or_time_increment"
error. The database library I'm using is called `better-sqlite3` and it
is synchronous. I was very new to Sqlite and this library. I thought if
it's synchronous then it may interfere with my timer so I moved it to a
subprocess. That is premature optimization right there. Though, I
learned a lot about child_process management with node but I think this
is the stupidest thing I've ever done. So, in this commit I've removed
it.